Finished level 1.Started working on level 2.Worked on economy and balancing stats

+++ Assets/Scripts/EnemyStats.cs
@@ -55,7 +55,7 @@ public class EnemyStats : MonoBehaviour
 
             if (!reachedEnd)
             {
-                CurrencyManager.Instance.AddGold((int)config.goldReward);
+                CurrencyManager.Instance.AddGold(config.goldReward);
                 Debug.Log("Am adaugat bani");
             }
             else
